Handle more corner cases for string vs hex
When a cpuset is created from a single number, it isn't always clear
whether the number is meant to be a hexmask, bitmask, or singleton
cpu set string.

The code always tries conversion from a string first (when a string
is provided), so for example "2" (or any other valid cpu) is converted
to a cpu_set_t with the bit for cpu2 set.

The trouble is when an *invalid* cpu id number is passed in, which is
also a valid hex or bitmask. In this case, we want to do the expected
thing, and fail for "99999" for example.

In order to do this, we use a now-amended heuristic:

 If the string begins with 0x, 00, or contains a string of digits longer
  than 7 chars followed by a comma, or cstr_to_cpuset() fails with
  an error other than E2BIG then assume we were passed a hexmask or
  bitmask.